# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "maklaGo API",
    "version": "1.0.0",
    "description": "Restaurant POS backend. **Tenant**: optional header `x-tenant-id` (slug or registry UUID). With `Authorization: Bearer` staff JWT, the tenant is resolved from the token when the header is omitted. **Language**: optional header **`lang`** (`fr`, `en`, `ar`; default **`fr`**) on every request — error messages and localized strings follow it; response includes **`Content-Language`**. **`POST /api/v1/auth/login`** accepts optional `x-tenant-id` for tenant-scoped login; without it, the normalized email must exist in registry **`staff_login_directory`** (slugless directory login). **`GET /api/v1/orders`** returns enriched catalog labels on lines (dashboard); **`GET /api/v1/mobile/orders`** uses a slimmer shell than **`POST /orders`**: no **`staff`**, no nested **`table`** object; dine-in adds root **`tableId`** + **`tableNumber`**. Lines add **`products[].name`**, **`extras[].name`**, optional **`extras[].typeId`**, optional **`compositionSnapshot`**. **`POST /api/v1/orders`** accepts **`products`** snapshots (`categoryId`, `id`, `count`, **`price`**, **`extras`**) for analytics; **`subtotal`**, **`tva`**, **`total`** use major currency units in JSON (e.g. 2.5); DB keeps cents internally (must match server); new orders start as **`waiting`**. **Platform** routes use `x-platform-key`. Errors: `{ \"error\": { \"code\", \"message\", \"details?\" } }`."
  },
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Staff JWT from `/api/v1/auth/login`."
      },
      "PlatformKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-platform-key",
        "description": "Must match `PLATFORM_API_KEY` on the server."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "waiting",
          "confirmed",
          "preparing",
          "completed",
          "canceled"
        ]
      },
...
      "EditOrderRequest": {
        "type": "object",
        "required": ["products", "subtotal", "tva", "total"],
        "description": "Replace entire cart on an open order. Same line/total rules as create (without orderType/tableId).",
        "properties": {
          "products": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["categoryId", "id", "count", "price"],
              "properties": {
                "categoryId": { "type": "string", "minLength": 12, "maxLength": 36 },
                "id": { "type": "string", "minLength": 12, "maxLength": 36 },
                "count": { "type": "integer", "minimum": 1, "maximum": 999 },
                "price": { "type": "number", "minimum": 0 },
                "extras": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["id", "count", "price"],
                    "properties": {
                      "id": { "type": "string" },
                      "count": { "type": "integer", "minimum": 0, "maximum": 999 },
                      "price": { "type": "number", "minimum": 0 }
                    }
                  }
                }
              }
            }
          },
          "subtotal": { "type": "number", "minimum": 0 },
          "tva": { "type": "number", "minimum": 0 },
          "total": { "type": "number", "minimum": 0 },
          "note": { "type": "string", "maxLength": 2000 },
          "customerName": { "type": "string", "maxLength": 200 },
          "discount": { "type": "number", "minimum": 0, "maximum": 100 }
        }
      },
...
      "OrderProductEnriched": {
        "type": "object",
        "description": "Backoffice order line: product/category labels and resolved extra names joined from catalog.",
        "properties": {
          "categoryId": { "type": "string", "minLength": 12, "maxLength": 36 },
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "kind": { "type": "string", "enum": ["simple", "composed"] },
          "categoryName": { "type": "string" },
          "image": { "type": "string", "nullable": true },
          "count": { "type": "integer" },
          "price": { "type": "number", "description": "Saved base unit price for this line (major currency units)." },
          "extras": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "count": { "type": "integer" },
                "price": { "type": "number" },
                "name": { "type": "string", "description": "Present when catalog extra exists." }
              }
            }
          },
          "compositionSnapshot": { "type": "object", "additionalProperties": true, "nullable": true }
        }
      },
      "OrderTableSerialized": {
        "type": "object",
        "description": "Table row embedded on dine-in orders. Display label is **`T` + `tableNumber`** (no separate name field). **`zone`** is null when the table has no zone.",
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "tableNumber": { "type": "integer", "minimum": 1 },
          "seatCount": { "type": "integer", "minimum": 1 },
          "zoneId": { "type": "string", "minLength": 12, "maxLength": 36, "nullable": true },
          "zone": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string", "minLength": 12, "maxLength": 36 },
              "name": { "type": "string" },
              "sortOrder": { "type": "integer" }
            }
          },
          "sortOrder": { "type": "integer" },
          "status": { "type": "string", "enum": ["free", "occupied"] },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "OrderSerializedEnriched": {
        "type": "object",
        "description": "Backoffice order (`GET /orders`, `GET /orders/{id}`, `PATCH …/status` HTTP body): enriched **`products`** (`OrderProductEnriched`). Shell matches **`OrderSerialized`** (table rules, `staff`, etc.).",
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "reference": {
            "type": "string",
            "description": "Unique ticket reference: **8** lowercase hexadecimal characters for newly created orders.",
            "example": "a3f9c21d"
          },
          "commandNumber": { "type": "integer" },
          "commandDate": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/OrderStatus" },
          "orderType": { "type": "string", "enum": ["dine_in", "takeaway"] },
          "tableId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "Present only for dine-in orders with an assigned table."
          },
          "staffId": { "type": "string", "minLength": 12, "maxLength": 36 },
          "note": { "type": "string", "nullable": true },
          "customerName": { "type": "string", "nullable": true },
          "discount": { "type": "number", "description": "Order-level discount percent (0–100)." },
          "discountPrice": { "type": "number" },
          "paymentMethod": { "type": "string", "enum": ["cash", "card", "unpaid"] },
          "subtotal": { "type": "number", "description": "Sum of line net totals (major currency units)." },
          "tva": { "type": "number", "description": "Sum of line TVA amounts (major currency units)." },
          "total": { "type": "number", "description": "Final total after discount (major currency units)." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "products": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderProductEnriched" }
          },
          "table": {
            "$ref": "#/components/schemas/OrderTableSerialized",
            "description": "Present only with `tableId` on dine-in orders when table row was loaded.",
            "nullable": true
          },
          "staff": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
...
      "ActivityAction": {
        "type": "string",
        "enum": [
          "order_created",
          "order_updated",
          "order_status_changed",
          "order_payment_recorded",
          "order_table_changed"
        ]
      },
      "ActivityEntityType": {
        "type": "string",
        "enum": ["order"]
      },
      "ActivityLogStaff": {
        "type": "object",
        "required": ["id", "firstName", "lastName", "email", "role"],
        "properties": {
          "id": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["owner", "manager", "cashier", "chef"] }
        }
      },
      "ActivityLog": {
        "type": "object",
        "required": ["id", "action", "entityType", "entityId", "createdAt", "staff"],
        "properties": {
          "id": { "type": "string" },
          "action": { "$ref": "#/components/schemas/ActivityAction" },
          "entityType": { "$ref": "#/components/schemas/ActivityEntityType" },
          "entityId": { "type": "string" },
          "summary": { "type": "string", "nullable": true },
          "metadata": { "type": "object", "additionalProperties": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "staff": { "$ref": "#/components/schemas/ActivityLogStaff" }
        }
      },
      "ActivityLogListResponse": {
        "type": "object",
        "required": ["logs"],
        "properties": {
          "logs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ActivityLog" }
          }
        }
      },
...
    },
    "parameters": {
      "OrderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 12,
          "maxLength": 36,
          "description": "Order public id (lowercase hex or legacy UUID)."
        }
      },
...
      "TenantHeaderOptional": {
        "name": "x-tenant-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Tenant slug (e.g. demo) or registry tenant UUID. Optional when the request includes a valid staff Bearer JWT—the server resolves tenant from the token."
      },
...
    }
  },
  "paths": {
...
    "/api/v1/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order by id",
        "description": "Returns **`order`** as **`OrderSerializedEnriched`** (display-ready lines).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          },
          {
            "$ref": "#/components/parameters/OrderId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": { "$ref": "#/components/schemas/OrderSerializedEnriched" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Orders"],
        "summary": "Edit order cart",
        "description": "Replace all lines on **confirmed** or **preparing** orders. Broadcasts **order.updated**. Resets **paymentMethod** to **unpaid** if previously paid.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/OrderId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditOrderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": { "$ref": "#/components/schemas/OrderSerializedEnriched" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation / not editable / totals mismatch" },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/api/v1/orders/{id}/logs": {
      "get": {
        "tags": ["Activity logs"],
        "summary": "Order activity history",
        "description": "Audit trail for one order (create, cart edit, status, payment, table move). Newest first.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/OrderId" },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityLogListResponse" }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity-logs": {
      "get": {
        "tags": ["Activity logs"],
        "summary": "List activity logs",
        "description": "Filter by **entityType**, **entityId**, and/or **action**. Defaults to latest 50 rows.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          {
            "name": "entityType",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ActivityEntityType" }
          },
          {
            "name": "entityId",
            "in": "query",
            "schema": { "type": "string", "minLength": 12, "maxLength": 36 }
          },
          {
            "name": "action",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ActivityAction" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityLogListResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity-logs/orders/{orderId}": {
      "get": {
        "tags": ["Activity logs"],
        "summary": "Order activity history (alias)",
        "description": "Same as **GET /api/v1/orders/{id}/logs**.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "minLength": 12, "maxLength": 36 }
          },
          {
            "name": "action",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ActivityAction" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityLogListResponse" }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
  }
}